Ejemplo de specifications

+++ java/spring-data-jpa/src/main/java/com/devutil/examples/spring/jpa/repository/customobjects/EntidadNombre.java
@@ -7,6 +7,6 @@ public class EntidadNombre {
 
     private final String nombre;
 
-    private final Integer contador;
+    private final Long contador;
 
 }
